refactor: move PeerCount into ream-p2p crate
