+++ crates/networking/p2p/Cargo.toml
@@ -16,6 +16,7 @@ asynchronous-codec = "0.7.0"
 delay_map.workspace = true
 discv5.workspace = true
 enr.workspace = true
+ethereum_serde_utils.workspace = true
 ethereum_ssz.workspace = true
 ethereum_ssz_derive.workspace = true
 futures.workspace = true
