+++ crates/networking/p2p/src/network/peer.rs
@@ -16,3 +16,15 @@ pub enum Direction {
     Outbound,
     Unknown,
 }
+
+#[derive(Default, Debug, Clone, Serialize)]
+pub struct PeerCount {
+    #[serde(with = "serde_utils::quoted_u64")]
+    pub disconnected: u64,
+    #[serde(with = "serde_utils::quoted_u64")]
+    pub connecting: u64,
+    #[serde(with = "serde_utils::quoted_u64")]
+    pub connected: u64,
+    #[serde(with = "serde_utils::quoted_u64")]
+    pub disconnecting: u64,
+}
