+++ crates/rpc/lean/Cargo.toml
@@ -22,4 +22,3 @@ ream-api-types-lean.workspace = true
 ream-chain-lean.workspace = true
 ream-consensus-lean.workspace = true
 ream-p2p.workspace = true
-ream-rpc-beacon.workspace = true
